:bug: fix (pulse): Fix import.
